+++ flink-kafka-processor/src/main/java/io/example/flink/model/Transaction.java
@@ -1,10 +1,6 @@
 package io.example.flink.model;
 
-import java.math.BigDecimal;
-
 import com.fasterxml.jackson.annotation.JsonProperty;
-
-
 
 
 /**
@@ -78,8 +74,8 @@ public class Transaction {
 		this.email = email;
 	}
 	
-	public BigDecimal getAmount(){
-		return new BigDecimal(this.transactionAmount);
+	public Float getAmount(){
+		return new Float(this.transactionAmount);
 	}
